requirements.txt
# Core scientific computing
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3

# Machine learning and optimization
scikit-learn==1.3.0
optuna==3.3.0

# Cryptography
pycryptodome==3.18.0
cryptography==41.0.3

# Blockchain and Web3
web3==6.8.0
py-solc-x==1.1.1

# Plotting and visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Data generation and statistics
faker==19.3.1
statsmodels==0.14.0

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Jupyter (optional, for interactive analysis)
jupyter==1.0.0
ipython==8.14.0
